[IOTDB-1453]Fix result set when the server query time filtered is And (#3452)
